chore: move useUniqueId to hooks directory
